feat(tss-2.1): delete store.ts — move Task type to check-helpers, inline fetch in check scripts
store.ts is no longer needed: agents use curl, and the two precheck scripts
(check-dev-task, check-deploy) now call the task store HTTP API via direct
fetch() calls using SHIPWRIGHT_TASK_STORE_URL + SHIPWRIGHT_TASK_STORE_TOKEN.

Task/TaskStatus types moved to check-helpers.ts and exported from there.
store.ts, store.unit.test.ts, and store.integration.test.ts deleted.
